+++ sdk/security/arm-security/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
     
-## 6.0.0-beta.4 (2023-04-11)
+## 6.0.0-beta.5 (2023-05-11)
     
 **Features**
 
@@ -16,6 +16,7 @@
   - Added operation group SecurityConnectorApplication
   - Added operation group SecurityConnectorApplications
   - Added operation group SecurityOperators
+  - Added operation group SensitivitySettings
   - Added operation Alerts.beginSimulate
   - Added operation Alerts.beginSimulateAndWait
   - Added operation Alerts.updateResourceGroupLevelStateToInProgress
@@ -26,6 +27,9 @@
   - Added Interface AdaptiveApplicationControlGroup
   - Added Interface AdaptiveNetworkHardening
   - Added Interface AdvancedThreatProtectionSetting
+  - Added Interface AggregationRequest
+  - Added Interface AggregationResponse
+  - Added Interface AggregationsOptionalParams
   - Added Interface Alert
   - Added Interface AlertPropertiesSupportingEvidence
   - Added Interface AlertSimulatorBundlesRequestProperties
@@ -71,6 +75,7 @@
   - Added Interface AzureDevOpsScopeEnvironmentData
   - Added Interface AzureResourceDetails
   - Added Interface AzureResourceIdentifier
+  - Added Interface BuiltInInfoType
   - Added Interface CefExternalSecuritySolution
   - Added Interface CefSolutionProperties
   - Added Interface Compliance
@@ -153,6 +158,11 @@
   - Added Interface GcpOrganizationalDataOrganization
   - Added Interface GcpProjectDetails
   - Added Interface GcpProjectEnvironmentData
+  - Added Interface GetSensitivitySettingsListResponse
+  - Added Interface GetSensitivitySettingsOptionalParams
+  - Added Interface GetSensitivitySettingsResponse
+  - Added Interface GetSensitivitySettingsResponseProperties
+  - Added Interface GetSensitivitySettingsResponsePropertiesMipInformation
   - Added Interface GithubScopeEnvironmentData
   - Added Interface GitlabScopeEnvironmentData
   - Added Interface GovernanceAssignment
@@ -191,6 +201,7 @@
   - Added Interface Identity
   - Added Interface InformationProtectionAwsOffering
   - Added Interface InformationProtectionPolicy
+  - Added Interface InfoType
   - Added Interface IngestionSetting
   - Added Interface IoTSecurityAggregatedAlert
   - Added Interface IoTSecurityAggregatedRecommendation
@@ -198,6 +209,7 @@
   - Added Interface IoTSecuritySolutionModel
   - Added Interface Issue
   - Added Interface JitNetworkAccessPolicy
+  - Added Interface Label
   - Added Interface ListCustomAlertRule
   - Added Interface LocalUserNotAllowed
   - Added Interface LogAnalyticsIdentifier
@@ -249,6 +261,7 @@
   - Added Interface SecuritySolutionsReferenceData
   - Added Interface SecuritySubAssessment
   - Added Interface SecurityTask
+  - Added Interface SensitivitySettingsListOptionalParams
   - Added Interface ServerVulnerabilityAssessment
   - Added Interface ServerVulnerabilityProperties
   - Added Interface Setting
@@ -262,7 +275,11 @@
   - Added Interface TwinUpdatesNotInAllowedRange
   - Added Interface UnauthorizedOperationsNotInAllowedRange
   - Added Interface UpdateIotSecuritySolutionData
+  - Added Interface UpdateSensitivitySettingsOptionalParams
+  - Added Interface UpdateSensitivitySettingsRequest
   - Added Interface WorkspaceSetting
+  - Added Type Alias AggregationsResponse
+  - Added Type Alias AggregationType
   - Added Type Alias APICollectionGetResponse
   - Added Type Alias APICollectionListNextResponse
   - Added Type Alias APICollectionListResponse
@@ -279,6 +296,7 @@
   - Added Type Alias EnvironmentDataUnion
   - Added Type Alias EnvironmentType
   - Added Type Alias GcpOrganizationalDataUnion
+  - Added Type Alias GetSensitivitySettingsOperationResponse
   - Added Type Alias GovernanceAssignmentsCreateOrUpdateResponse
   - Added Type Alias GovernanceAssignmentsGetResponse
   - Added Type Alias GovernanceAssignmentsListNextResponse
@@ -298,6 +316,7 @@
   - Added Type Alias HealthReportsListResponse
   - Added Type Alias IsEnabled
   - Added Type Alias MinimalSeverity
+  - Added Type Alias MipIntegrationStatus
   - Added Type Alias OperationResult
   - Added Type Alias Roles
   - Added Type Alias ScanningMode
@@ -309,15 +328,18 @@
   - Added Type Alias SecurityOperatorsCreateOrUpdateResponse
   - Added Type Alias SecurityOperatorsGetResponse
   - Added Type Alias SecurityOperatorsListResponse
+  - Added Type Alias SensitivitySettingsListResponse
   - Added Type Alias StatusName
   - Added Type Alias SubPlan
   - Added Type Alias Type
+  - Added Type Alias UpdateSensitivitySettingsResponse
   - Interface AlertsSimulateOptionalParams has a new optional parameter resumeFrom
   - Interface AlertsSimulateOptionalParams has a new optional parameter updateIntervalInMs
   - Interface DefenderForServersAwsOfferingArcAutoProvisioning has a new optional parameter cloudRoleArn
   - Interface DefenderForServersAwsOfferingArcAutoProvisioning has a new optional parameter configuration
   - Interface ScanProperties has a new optional parameter lastScanTime
   - Type of parameter offeringType of interface CloudOffering is changed from "CspmMonitorAws" | "DefenderForContainersAws" | "DefenderForServersAws" | "InformationProtectionAws" to "CspmMonitorAws" | "DefenderForContainersAws" | "DefenderForServersAws" | "DefenderForDatabasesAws" | "InformationProtectionAws" | "CspmMonitorGcp" | "DefenderForServersGcp" | "DefenderForDatabasesGcp" | "DefenderForContainersGcp" | "CspmMonitorGithub" | "CspmMonitorAzureDevOps" | "DefenderCspmAws" | "DefenderCspmGcp" | "DefenderForDevOpsGithub" | "DefenderForDevOpsAzureDevOps" | "CspmMonitorGitLab" | "DefenderForDevOpsGitLab"
+  - Added Enum KnownAggregationType
   - Added Enum KnownApplicationConditionOperator
   - Added Enum KnownApplicationSourceResourceType
   - Added Enum KnownCode
@@ -328,6 +350,7 @@
   - Added Enum KnownGovernanceRuleType
   - Added Enum KnownIsEnabled
   - Added Enum KnownMinimalSeverity
+  - Added Enum KnownMipIntegrationStatus
   - Added Enum KnownOperationResult
   - Added Enum KnownRoles
   - Added Enum KnownScanningMode